+++ examples/simple/src/schema/__global__schema.ts
@@ -10,6 +10,7 @@ import {
   UUIDType,
   StringType,
   TimestampType,
+  UUIDListType,
 } from "@snowtop/ent/schema/";
 
 const glo: GlobalSchema = {
@@ -40,6 +41,18 @@ const glo: GlobalSchema = {
           tsType: "NotifType",
           graphQLType: "NotifType",
         }),
+        homeAddressId: UUIDType({
+          nullable: true,
+          fieldEdge: {
+            schema: "Address",
+          },
+        }),
+        allAddressIds: UUIDListType({
+          fieldEdge: {
+            schema: "Address",
+          },
+          nullable: true,
+        }),
       },
     }),
     responses: EnumType({
@@ -63,6 +76,19 @@ const glo: GlobalSchema = {
         date: TimestampType(),
         phone_number: PhoneNumberType({ nullable: true }),
         email_address: EmailType({ nullable: true }),
+        creator_id: UUIDType({
+          polymorphic: {
+            types: ["User"],
+          },
+          nullable: true,
+        }),
+        // doesn't seem like we support this...
+        // owner_ids: UUIDListType({
+        //   polymorphic: {
+        //     types: ["User"],
+        //   },
+        //   nullable: true,
+        // }),
       },
     }),
   },