+++ internal/field/field_info.go
@@ -11,8 +11,9 @@ import (
 )
 
 type Options struct {
-	SortFields     bool
-	FieldOverrides map[string]*input.FieldOverride
+	SortFields              bool
+	FieldOverrides          map[string]*input.FieldOverride
+	ForceDisableBuilderType bool
 }
 
 // NewFieldInfoFromInputs generates Fields based on FieldInputs
@@ -50,6 +51,9 @@ func NewFieldInfoFromInputs(cfg codegenapi.Config, nodeName string, fields []*in
 		}
 		if err := fieldInfo.addField(f); err != nil {
 			errs = append(errs, err)
+		}
+		if options.ForceDisableBuilderType {
+			f.disableBuilderType = true
 		}
 		for _, derivedField := range field.DerivedFields {
 			f2, err := newFieldFromInput(cfg, nodeName, derivedField)
@@ -93,7 +97,6 @@ type FieldInfo struct {
 	NonEntFields []*NonEntField
 	// keep track of computed fields just to know they exist
 	computedFields map[string]bool
-	getFieldsFn    bool
 
 	primaryKeys []string