+++ release_image/main.go
@@ -18,7 +18,7 @@ import (
 )
 
 // next tag to use
-const TAG = "v0.2.0-alpha.4"
+const TAG = "v0.2.0-alpha.11"
 
 // current node gets latest tag...
 const CURRENT_NODE_VERSION = 18
@@ -27,13 +27,13 @@ const REPO = "ghcr.io/lolopinto/ent"
 const UPDATE_LATEST = true
 
 var NODE_VERSIONS = []int{
-	// 16,
-	18,
-	20,
+	// 18,
+	// 20,
+	22,
 }
 
 const AUTO_SCHEMA_VERSION = "0.0.32"
-const TSENT_VERSION = "v0.2.0-alpha.4"
+const TSENT_VERSION = "v0.2.0-alpha.11"
 
 // TODO release notes